Fixes for Windows compilation
Update PortableStdint to a later version

Fix List_removeLast_ implementation

(This function was unused in the runtime but it was incorrect
as size_t is unsigned)

Add missing definitions for Windows

Give function typedef the correct typedef

(MSVC seems to return garbage if a function returning bool
is cast to a function returning int.)

Fix compiler warnings

io_calloc could potentially return NULL

Fix bugs in IO_CHECK_ALLOC

+++ libs/basekit/source/PortableStrptime.c
@@ -114,7 +114,7 @@ char *io_strptime(char *buf, char *fmt, struct tm *tm) {
                            "%A, %B, %e, %Y"};
 
     char c, *ptr;
-    long i, len;
+    size_t i, len;
 
     len = 0;
     ptr = fmt;
